requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.7
colorama==0.4.6
Django==6.0.4
django-browser-reload==1.21.0
iniconfig==2.3.0
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
psycopg2-binary==2.9.12
Pygments==2.20.0
pytest==9.0.3
pytest-django==4.12.0
python-dotenv==1.2.2
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2026.1
requests==2.32.3
sweetalert2-django==0.1.0
